Allow for large number of add-ons.
As there are now a lot of Debug Bar add-ons out there, the panel list on the left might run out of screen estate.
The tweaks in this PR allow for a scroll bar to be added to the panel list and make the list a little more compact.

Also fixes an issue where the panel menu links wouldn't be full-width if the front-end theme changed the `box-sizing`.