Don't prevent error logging when silencing error notices.
Any existing error log directives should be respected.
By silencing the error notices, `trigger_error()` was completely bypassed which meant that error did not show on screen nor in the logs and as the admin bar highlight wasn't working nearly as well as it should, this meant that people were missing errors.