Don't show E_USER_DEPRECATED in the PHP Panel as they should show in the Deprecated panel.
This removes duplicates of these notices from the PHP Panel. However it is presumed that devs use the appropriate WP `_deprecated_...()` function call.

Removing this constant here will  mean that when an error is throw with this level without using the WP function, it will not show up at all.