Finish off code style fixes.
I noticed that the last commits before the 0.9 release contained quite some code style fixes.
As the code is now close to complying with the WP Extra rules, we might as well go all the way.

This PR contains the minimal changes needed to comply with the `WordPress-Extra` WPCS ruleset with the exception of two rules which will be so annotated in the ruleset (to be pulled in a separate PR).