Fix warning when php_uname() is disabled.
PHP provides an ini_directive through which hosts can disable functions for security reasons.
`disable_functions = php_uname`
Recently a user reported the `php_uname()` function to be disabled causing a warning to that effect.

This "fix" is a work-around. If `php_uname` is enabled, it will use the host name as provided thorugh that function, otherwise it will check the `$_SERVER['SERVER_NAME']` value and if all else fails, it will just display the text string 'Site'.

Fixes https://wordpress.org/support/topic/php_uname-has-been-disabled-for-security-reasons-2/