Add visibility modifiers to all methods and properties
Some new checks have been added to WPCS to verify that visibility modifiers are always used which will cause the builds for this plugin to break with the current codebase.

Well, we know those had to be added to this plugin anyhow, so now is as good a time as any to do this.

As a general rule of thumb I've made everything `protected`, except:
* Class constructors - these need to be `public`.
* The error handler being registered with PHP - this has to be a `public` method.
* Methods which are being called via hooks, including `Debug_Bar_Panel::debug_bar_classes()` - these need to be `public`.
* Methods which are being called from the main `Debug_Bar` class - i.e. `Debug_Bar_Panel::is_visible()`, `Debug_Bar_Panel::title()`, `Debug_Bar_Panel::prerender()` and `Debug_Bar_Panel::render()` - these need to be `public`.

Using `protected` whenever possible creates the right balance as:
* Add-ons will need to overload certain methods from the `Debug_Bar_Panel` class, which means that using `private` is not an option for most methods.
* Also, using `protected` will not break compatibility for add-on plugins which already have a visibility defined as overloaded methods can be "more" visible, not less.

I've also renamed one method which was indicated in the PHP 4 manner to be private with an underscore prefix.