Start logging of notices early. (#38)
The way things currently were, logging would not start until the `admin_bar_init` hook was reached. That meant that any PHP and/or Deprecated notices before that point would be missed.

This PR fixes that by starting the logging as soon as the Debug Bar class is initialized.

In case later - in the actual `init()` functions - it is detected that the Debug Bar will not be displayed, the logging will be stopped again.