PHPCS ruleset: minor updates for the upcoming WPCS 1.0.0 release
As the Travis build uses the `develop` branch of WPCS, the sniff name changes which will be contained in the WPCS 1.0.0 release are already in effect.

Also, WPCS 0.14.0 offered a way to simplify the ruleset a little by only needing to set the minimum supported WP version once, so let's use that.